[vcpkg] Restore layering and purge unused ParagraphParseResult
